feat: clean up code and add LOCMAF test asset files

+++ internal/sub/sub.go
@@ -678,7 +678,7 @@ func parseLocmafInit(track *internal.Track) (*mp4.InitSegment, error) {
 		return nil, fmt.Errorf("invalid locmaf init payload length")
 	}
 	pos += n
-	if headerID != internal.MoovHeader {
+	if headerID != uint64(internal.MoovHeader) {
 		return nil, fmt.Errorf("unsupported locmaf init header %d", headerID)
 	}
 	if locPayloadLength > uint64(len(initDataBytes)-pos) {
